# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/port/berths": {
      "get": {
        "tags": [
          "port"
        ],
        "summary": "List Berths",
        "operationId": "list_berths_api_v1_port_berths_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortBerthResponse"
                  },
                  "type": "array",
                  "title": "Response List Berths Api V1 Port Berths Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/port/sync-status": {
      "get": {
        "tags": [
          "port"
        ],
        "summary": "Port Sync Status",
        "operationId": "port_sync_status_api_v1_port_sync_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "port_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "title": "Port Id"
            }
          },
          {
            "name": "vessel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "title": "Vessel Id"
            }
          },
          {
            "name": "train_arrival_hours",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 720,
              "exclusiveMinimum": 0,
              "title": "Train Arrival Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortSyncStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/predictive/delay": {
      "post": {
        "tags": [
          "predictive"
        ],
        "summary": "Predict Route Delay",
        "operationId": "predict_route_delay_api_v1_predictive_delay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelayPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelayPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/predictive/dust-risk": {
      "get": {
        "tags": [
          "predictive"
        ],
        "summary": "Get Dust Storm Risk",
        "operationId": "get_dust_storm_risk_api_v1_predictive_dust_risk_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "Latitude",
              "title": "Lat"
            },
            "description": "Latitude"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "Longitude",
              "title": "Lon"
            },
            "description": "Longitude"
          },
          {
            "name": "location",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "Node name",
              "title": "Location"
            },
            "description": "Node name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DustStormRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DelayPredictionRequest": {
        "properties": {
          "route_id": {
            "type": "string",
            "format": "uuid",
            "title": "Route Id"
          }
        },
        "type": "object",
        "required": [
          "route_id"
        ],
        "title": "DelayPredictionRequest"
      },
      "DelayPredictionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "const": "STORED_ROUTE_FEATURES",
            "title": "Source",
            "default": "STORED_ROUTE_FEATURES"
          },
          "route_id": {
            "type": "string",
            "format": "uuid",
            "title": "Route Id"
          },
          "predicted_delay_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Delay Minutes"
          },
          "confidence_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Pct"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "LOW",
                  "MODERATE",
                  "HIGH",
                  "CRITICAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "primary_bottleneck_segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Bottleneck Segment"
          },
          "weather_impact_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Impact Pct"
          },
          "congestion_impact_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Congestion Impact Pct"
          },
          "optimal_dispatch_window": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimal Dispatch Window"
          },
          "limitation": {
            "type": "string",
            "title": "Limitation"
          }
        },
        "type": "object",
        "required": [
          "status",
          "route_id",
          "predicted_delay_minutes",
          "confidence_pct",
          "risk_level",
          "weather_impact_pct",
          "congestion_impact_pct",
          "optimal_dispatch_window",
          "limitation"
        ],
        "title": "DelayPredictionResponse"
      },
...
      "DustStormRiskResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "dust_risk_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dust Risk Index"
          },
          "airborne_particulate_pm10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Airborne Particulate Pm10"
          },
          "visibility_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility Km"
          },
          "warning_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "SAFE",
                  "ADVISORY",
                  "SEVERE",
                  "HAZARD"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning Level"
          },
          "recommended_speed_limit_kmh": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Speed Limit Kmh"
          },
          "operational_limitation": {
            "type": "string",
            "title": "Operational Limitation"
          }
        },
        "type": "object",
        "required": [
          "status",
          "source",
          "location",
          "lat",
          "lon",
          "observed_at",
          "fetched_at",
          "dust_risk_index",
          "airborne_particulate_pm10",
          "visibility_km",
          "warning_level",
          "recommended_speed_limit_kmh",
          "operational_limitation"
        ],
        "title": "DustStormRiskResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortBerthResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "port_name": {
            "type": "string",
            "title": "Port Name"
          },
          "berth_identifier": {
            "type": "string",
            "title": "Berth Identifier"
          },
          "vessel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Name"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "port_name",
          "berth_identifier",
          "vessel_name",
          "window_start",
          "window_end"
        ],
        "title": "PortBerthResponse"
      },
...
      "PortSyncStatus": {
        "properties": {
          "aligned": {
            "type": "boolean",
            "title": "Aligned"
          },
          "berth_id": {
            "type": "string",
            "title": "Berth Id"
          },
          "vessel_status": {
            "type": "string",
            "title": "Vessel Status"
          },
          "loading_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Loading Window Start"
          },
          "loading_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Loading Window End"
          },
          "sync_score": {
            "type": "number",
            "title": "Sync Score"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "type": "object",
        "required": [
          "aligned",
          "berth_id",
          "vessel_status",
          "loading_window_start",
          "loading_window_end",
          "sync_score"
        ],
        "title": "PortSyncStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}